@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import App from './components/wrappers/App';
 
 // Add bootstrap
@@ -8,7 +8,16 @@ import 'bootstrap/dist/css/bootstrap.css';
 // Add our style
 import './assets/style/index.css';
 
-ReactDOM.render(
-    <App/>,
-    document.getElementById('root')
+// ReactDOM.render(
+//     <App/>,
+//     document.getElementById('root')
+// );
+
+const rootElement = document.getElementById('root');
+
+const root = ReactDOM.createRoot(rootElement);
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
 );
