@@ -2,17 +2,19 @@ import React from 'react';
 import CheckBox from './CheckBox';
 
 export default function TodoItem(props) {
-    const {data, changeStatus} = props;
-    const handleChange = (checked) => changeStatus(data.id, checked);
-    const className = 'todo-item ui-state-default ' + (data.completed === true ? 'completed' : 'pending');
+  const { data, changeStatus } = props;
 
-    return (
-        <li className={className}>
-            <div className="checkbox">
-                <label>
-                    <CheckBox checked={data.completed} onChange={handleChange}/> {data.text}
-                </label>
-            </div>
-        </li>
-    );
+  return (
+    <li className={`todo-item ui-state-default ${data.completed ? 'completed' : 'pending'}`}>
+      <div className="checkbox">
+        <label>
+          <CheckBox
+            checked={data.completed}
+            onChange={(checked) => changeStatus(data.id, checked)} // Trigger status change with confirmation on completion
+          />
+          {data.text}
+        </label>
+      </div>
+    </li>
+  );
 }
